requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==1.17.1
colorama==0.4.6
cssselect2==0.8.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.3.0
fonttools==4.58.5
gunicorn==21.2.0
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
pytest==8.4.1
pytest-django==4.11.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
requests==2.31.0
python-decouple==3.8
debugpy==1.8.1
python-docx==1.2.0

# New dependencies for retainer app
pandas==1.5.3
openpyxl==3.1.2
celery==5.3.4
redis==5.0.1
numpy==1.24.3